code refactor into three files. update readme.
